Refactor toast components for manage multiple toasts
- Update Toaster component to manage multiple toasts using the new structure
- Replace the old toast store implementation with a new Toast component
